# packages/api-types/openapi.json
{
  "components": {
    "schemas": {
      "AjustesDeAgenda": {
        "description": "Los números que el brief deja configurables (AGD-1, RSV-4, D10, REV-1).\n\n**Ninguno vive como constante en el código.** Cambiar la granularidad a 30 minutos o la\nventana de cancelación a 24 horas es editar esta fila, no desplegar.",
        "properties": {
          "antelacion_maxima_dias": {
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Antelacion Maxima Dias",
            "type": "integer"
          },
          "antelacion_minima_minutos": {
            "maximum": 43200.0,
            "minimum": 0.0,
            "title": "Antelacion Minima Minutos",
            "type": "integer"
          },
          "auto_confirmar": {
            "title": "Auto Confirmar",
            "type": "boolean"
          },
          "bloquear_tras_ausencias": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "NULL = no bloquear a nadie por no-shows (RSV-5)",
            "title": "Bloquear Tras Ausencias"
          },
          "granularidad_minutos": {
            "maximum": 120.0,
            "minimum": 5.0,
            "title": "Granularidad Minutos",
            "type": "integer"
          },
          "permitir_cualquier_profesional": {
            "title": "Permitir Cualquier Profesional",
            "type": "boolean"
          },
          "resumen_diario": {
            "title": "Resumen Diario",
            "type": "boolean"
          },
          "ventana_cancelacion_horas": {
            "maximum": 336.0,
            "minimum": 0.0,
            "title": "Ventana Cancelacion Horas",
            "type": "integer"
          },
          "ventana_resena_dias": {
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Ventana Resena Dias",
            "type": "integer"
          }
        },
        "required": [
          "granularidad_minutos",
          "antelacion_minima_minutos",
          "antelacion_maxima_dias",
          "auto_confirmar",
          "ventana_cancelacion_horas",
          "ventana_resena_dias",
          "permitir_cualquier_profesional",
          "resumen_diario"
        ],
        "title": "AjustesDeAgenda",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Reservas y marketplace de belleza y bienestar en Panamá. Cada endpoint cita el requisito del brief que cubre.",
    "title": "API de Bukeo",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/negocio/ajustes": {
      "get": {
        "operationId": "leer_ajustes_api_v1_negocio_ajustes_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AjustesDeAgenda"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Los números configurables del salón (AGD-1, D10)",
        "tags": [
          "agenda del negocio"
        ]
      },
      "patch": {
        "description": "Cambia lo que venga. **No toca las citas que ya existen.**\n\nSubir la antelación mínima a tres horas no cancela la cita de dentro de dos: esa ya está\nhecha y el hueco está apartado. Lo que cambia es lo que se ofrece a partir de ahora.",
        "operationId": "cambiar_ajustes_api_v1_negocio_ajustes_patch",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambioDeAjustes"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AjustesDeAgenda"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cambiar los números sin desplegar (AGD-1, RSV-4, D10)",
        "tags": [
          "agenda del negocio"
        ]
      }
    },
...
    "/api/v1/negocio/ausencias": {
      "get": {
        "description": "Los bloqueos del equipo en un rango. Por defecto, de hoy en adelante.",
        "operationId": "listar_ausencias_api_v1_negocio_ausencias_get",
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Desde"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hasta"
            }
          },
          {
            "in": "query",
            "name": "incluir_levantadas",
            "required": false,
            "schema": {
              "default": false,
              "title": "Incluir Levantadas",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Ausencia"
                  },
                  "title": "Response Listar Ausencias Api V1 Negocio Ausencias Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vacaciones, días libres y cierres puntuales (AGD-3)",
        "tags": [
          "agenda del negocio"
        ]
      },
      "post": {
        "description": "Bloquea a una persona o a **todo el equipo**, que es como se cierra el salón un día.\n\nDevuelve una lista porque un cierre del salón son tantas filas como profesionales activos:\nla agenda es por persona y la restricción de exclusión también, así que no hay una fila\n«del negocio» que pueda bloquear a todos a la vez. Verlo como lista es además lo honesto,\nporque después se puede levantar el bloqueo de uno sin levantar el de los demás.\n\nSi dentro del tramo ya hay citas, **no se bloquea nada** y sale `409`: la transacción es\nuna sola, así que o entran todos los bloqueos o no entra ninguno.",
        "operationId": "crear_ausencia_api_v1_negocio_ausencias_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AltaDeAusencia"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Ausencia"
                  },
                  "title": "Response Crear Ausencia Api V1 Negocio Ausencias Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bloquear un tramo (AGD-3, AGD-6)",
        "tags": [
          "agenda del negocio"
        ]
      }
    },
    "/api/v1/negocio/ausencias/{ausencia_id}": {
      "delete": {
        "description": "Levanta el bloqueo **sin borrar la fila**, igual que cancelar una cita.\n\nLa restricción de exclusión solo mira los bloqueos `activo`, así que el hueco vuelve a\nofrecerse de inmediato; y queda el rastro de que ese día estuvo cerrado, que es lo que\nhace falta para entender una agenda vacía tres meses después.",
        "operationId": "levantar_ausencia_api_v1_negocio_ausencias__ausencia_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "ausencia_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ausencia Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ausencia"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Levantar un bloqueo (AGD-3)",
        "tags": [
          "agenda del negocio"
        ]
      }
    },
...
    "/api/v1/negocio/feriados": {
      "get": {
        "description": "Los feriados del país del negocio, con la marca de si el salón ya cerró ese día.\n\n**Sugeridos, no impuestos**: la respuesta dice qué días son feriado y cuáles ya están\ncerrados; cerrar es un `POST /negocio/ausencias` como cualquier otro. Media ciudad abre el\ndía de la madre.",
        "operationId": "listar_feriados_api_v1_negocio_feriados_get",
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Desde"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hasta"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeriadoSugerido"
                  },
                  "title": "Response Listar Feriados Api V1 Negocio Feriados Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feriados de Panamá sugeridos (AGD-6)",
        "tags": [
          "agenda del negocio"
        ]
      }
    },
...
    "/api/v1/negocio/horario": {
      "get": {
        "operationId": "leer_horario_api_v1_negocio_horario_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TramoDelNegocio"
                  },
                  "title": "Response Leer Horario Api V1 Negocio Horario Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Horario semanal del salón (AGD-2, NEG-1)",
        "tags": [
          "agenda del negocio"
        ]
      },
      "put": {
        "description": "Reemplaza el horario entero.\n\n**No borra ni cancela las citas que queden fuera** del horario nuevo (caso 4 del motor):\nel negocio decide qué hacer con ellas. Que un cambio de horario cancelara citas en silencio\nsería la peor sorpresa posible un lunes por la mañana.",
        "operationId": "poner_horario_api_v1_negocio_horario_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/HorarioDelDia"
                },
                "title": "Horario",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HorarioDelDia"
                  },
                  "title": "Response Poner Horario Api V1 Negocio Horario Put",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Horario semanal del negocio (ONB-2)",
        "tags": [
          "alta de negocio"
        ]
      }
    },
...
  }
}